Move cursor icon and event routing control to source
