+++ src/server/frontend_wayland/wl_data_source.cpp
@@ -15,8 +15,14 @@
  */
 
 #include "wl_data_source.h"
-#include "mir/scene/clipboard.h"
+
 #include "mir/executor.h"
+#include "mir/frontend/drag_icon_controller.h"
+#include "mir/frontend/pointer_input_dispatcher.h"
+#include "mir/scene/clipboard.h"
+#include "mir/scene/session.h"
+#include "mir/scene/surface.h"
+#include "mir/shell/surface_specification.h"
 #include "mir/wayland/weak.h"
 
 #include <vector>
@@ -24,6 +30,7 @@
 namespace mf = mir::frontend;
 namespace ms = mir::scene;
 namespace mw = mir::wayland;
+using namespace mir::geometry;
 
 class mf::WlDataSource::ClipboardObserver : public ms::ClipboardObserver
 {
@@ -60,14 +67,65 @@ private:
     wayland::Weak<WlDataSource> const owner;
 };
 
+mf::WlDataSource::DragIconSurface::DragIconSurface(WlSurface* icon, std::shared_ptr<DragIconController> drag_icon_controller)
+    : NullWlSurfaceRole(icon),
+      surface{icon},
+      drag_icon_controller{std::move(drag_icon_controller)}
+{
+    icon->set_role(this);
+
+    auto spec = shell::SurfaceSpecification();
+    spec.width = surface.value().buffer_size()->width;
+    spec.height = surface.value().buffer_size()->height;
+    spec.streams = std::vector<shell::StreamSpecification>{};
+    spec.input_shape = std::vector<Rectangle>{};
+    spec.depth_layer = mir_depth_layer_overlay;
+
+    surface.value().populate_surface_data(spec.streams.value(), spec.input_shape.value(), {});
+
+    auto const& session = surface.value().session;
+
+    shared_scene_surface =
+        session->create_surface(session, wayland::Weak<WlSurface>(surface), spec, nullptr, nullptr);
+
+    DragIconSurface::drag_icon_controller->set_drag_icon(shared_scene_surface);
+}
+
+mf::WlDataSource::DragIconSurface::~DragIconSurface()
+{
+    if (surface)
+    {
+        surface.value().clear_role();
+
+        if (shared_scene_surface)
+        {
+            auto const& session = surface.value().session;
+            session->destroy_surface(shared_scene_surface);
+        }
+    }
+}
+
+auto mf::WlDataSource::DragIconSurface::scene_surface() const -> std::optional<std::shared_ptr<scene::Surface>>
+{
+    return shared_scene_surface;
+}
+
 class mf::WlDataSource::Source : public ms::DataExchangeSource
 {
 public:
-    Source(WlDataSource& wl_data_source)
-        : wayland_executor{wl_data_source.wayland_executor},
-          wl_data_source{&wl_data_source},
-          types{wl_data_source.mime_types}
+    Source(WlDataSource& wl_data_source) :
+        wayland_executor{wl_data_source.wayland_executor},
+        wl_data_source{&wl_data_source},
+        types{wl_data_source.mime_types}
     {
+    }
+
+    ~Source()
+    {
+        if (wl_data_source)
+        {
+            wl_data_source.value().end_drag_n_drop_gesture();
+        }
     }
 
     auto mime_types() const -> std::vector<std::string> const& override
@@ -91,6 +149,7 @@ public:
         if (wl_data_source)
         {
             wl_data_source.value().send_cancelled_event();
+            wl_data_source.value().end_drag_n_drop_gesture();
         }
     }
 
@@ -98,7 +157,8 @@ public:
     {
         if (wl_data_source)
         {
-            return wl_data_source.value().send_dnd_drop_performed_event_if_supported();
+            wl_data_source.value().send_dnd_drop_performed_event_if_supported();
+            wl_data_source.value().end_drag_n_drop_gesture();
         }
     }
 
@@ -123,11 +183,15 @@ private:
 mf::WlDataSource::WlDataSource(
     wl_resource* new_resource,
     std::shared_ptr<Executor> const& wayland_executor,
-    scene::Clipboard& clipboard)
+    scene::Clipboard& clipboard,
+    std::shared_ptr<DragIconController> drag_icon_controller,
+    std::shared_ptr<PointerInputDispatcher> pointer_input_dispatcher)
     : mw::DataSource{new_resource, Version<3>()},
       wayland_executor{wayland_executor},
       clipboard{clipboard},
-      clipboard_observer{std::make_shared<ClipboardObserver>(this)}
+      clipboard_observer{std::make_shared<ClipboardObserver>(this)},
+      drag_icon_controller{std::move(drag_icon_controller)},
+      pointer_input_dispatcher{std::move(pointer_input_dispatcher)}
 {
     clipboard.register_interest(clipboard_observer, *wayland_executor);
 }
@@ -158,12 +222,20 @@ void mf::WlDataSource::set_clipboard_paste_source()
     clipboard.set_paste_source(source);
 }
 
-void mf::WlDataSource::set_drag_n_drop_source()
+void mf::WlDataSource::start_drag_n_drop_gesture(std::optional<wl_resource*> const& icon)
 {
     send_target_event(std::nullopt);
     auto const source = std::make_shared<Source>(*this);
     dnd_source = source;
     clipboard.set_drag_n_drop_source(source);
+    pointer_input_dispatcher->start_ignore_gesture_owner();
+
+    if (icon)
+    {
+        auto const icon_surface = WlSurface::from(icon.value());
+
+        drag_surface.emplace(icon_surface, drag_icon_controller);
+    }
 }
 
 void mf::WlDataSource::offer(std::string const& mime_type)
@@ -214,4 +286,14 @@ void mf::WlDataSource::drag_n_drop_source_cleared(std::shared_ptr<scene::DataExc
 void mir::frontend::WlDataSource::set_actions(uint32_t dnd_actions)
 {
     this->dnd_actions = dnd_actions;
+}
+
+void mir::frontend::WlDataSource::end_drag_n_drop_gesture()
+{
+    pointer_input_dispatcher->end_ignore_gesture_owner();
+    drag_surface.reset();
+    if (auto const source = dnd_source.lock())
+    {
+        clipboard.clear_drag_n_drop_source(source);
+    }
 }
